refactor: adjust file structure
release-as: 0.32.9
