+++ packages/editor/src/extensions/inline/index.ts
@@ -1,3 +1,4 @@
 export * from "./inline-deco-plugin"
-export * from "./define"
+export * from "./inline-mark-extensions"
 export * from "./inline-mark-plugin"
+export * from "./inline-types"
