+++ packages/home/src/styles/cache.ts
@@ -1,5 +1,5 @@
 import createCache from "@emotion/cache"
 
 export function createEmotionCache() {
-    return createCache({ key: "css",prepend: true, })
+    return createCache({ key: "css", prepend: true })
 }
